+++ src/jdk.jfr/share/classes/jdk/jfr/EventSettings.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2016, 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -26,9 +26,15 @@
 package jdk.jfr;
 
 import java.time.Duration;
+import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 
+import jdk.jfr.internal.PlatformEventType;
+import jdk.jfr.internal.PlatformRecorder;
+import jdk.jfr.internal.PlatformRecording;
+import jdk.jfr.internal.PrivateAccess;
+import jdk.jfr.internal.Type;
 import jdk.jfr.internal.management.EventSettingsModifier;
 
 /**
@@ -45,6 +51,115 @@ import jdk.jfr.internal.management.EventSettingsModifier;
  * @since 9
  */
 public abstract class EventSettings {
+
+    // Purpose of InternalAccess is to give classes in jdk.jfr.internal
+    // access to package private methods in this package (jdk.jfr).
+    //
+    // The initialization could be done in any class in this package,
+    // but this one was chosen because it is lightweight.
+    static {
+        PrivateAccess.setPrivateAccess(new InternalAccess());
+    }
+
+    private static final class InternalAccess extends PrivateAccess {
+
+        @Override
+        public Type getType(Object o) {
+            if (o instanceof AnnotationElement ae) {
+                return ae.getType();
+            }
+            if (o instanceof EventType et) {
+                return et.getType();
+            }
+            if (o instanceof ValueDescriptor vd) {
+                return vd.getType();
+            }
+            if (o instanceof SettingDescriptor sd) {
+                return sd.getType();
+            }
+            throw new Error("Unknown type " + o.getClass());
+        }
+
+        @Override
+        public Configuration newConfiguration(String name, String label, String description, String provider, Map<String, String> settings, String contents) {
+            return new Configuration(name, label, description, provider, settings, contents);
+        }
+
+        @Override
+        public EventType newEventType(PlatformEventType platformEventType) {
+            return new EventType(platformEventType);
+        }
+
+        @Override
+        public AnnotationElement newAnnotation(Type annotationType, List<Object> values, boolean boot) {
+            return new AnnotationElement(annotationType, values, boot);
+        }
+
+        @Override
+        public ValueDescriptor newValueDescriptor(String name, Type fieldType, List<AnnotationElement> annos, int dimension, boolean constantPool, String fieldName) {
+            return new ValueDescriptor(fieldType, name, annos, dimension, constantPool, fieldName);
+        }
+
+        @Override
+        public PlatformRecording getPlatformRecording(Recording r) {
+            return r.getInternal();
+        }
+
+        @Override
+        public PlatformEventType getPlatformEventType(EventType eventType) {
+            return eventType.getPlatformEventType();
+        }
+
+        @Override
+        public boolean isConstantPool(ValueDescriptor v) {
+            return v.isConstantPool();
+        }
+
+        @Override
+        public void setAnnotations(ValueDescriptor v, List<AnnotationElement> a) {
+            v.setAnnotations(a);
+        }
+
+        @Override
+        public void setAnnotations(SettingDescriptor s, List<AnnotationElement> a) {
+           s.setAnnotations(a);
+        }
+
+        @Override
+        public String getFieldName(ValueDescriptor v) {
+            return v.getJavaFieldName();
+        }
+
+        @Override
+        public ValueDescriptor newValueDescriptor(Class<?> type, String name) {
+            return new ValueDescriptor(type, name, List.of(), true);
+        }
+
+        @Override
+        public SettingDescriptor newSettingDescriptor(Type type, String name, String defaultValue, List<AnnotationElement> annotations) {
+            return new SettingDescriptor(type, name, defaultValue, annotations);
+        }
+
+        @Override
+        public boolean isUnsigned(ValueDescriptor v) {
+            return v.isUnsigned();
+        }
+
+        @Override
+        public PlatformRecorder getPlatformRecorder() {
+            return FlightRecorder.getFlightRecorder().getInternal();
+        }
+
+        @Override
+        public EventSettings newEventSettings(EventSettingsModifier esm) {
+            return new EventSettings.DelegatedEventSettings(esm);
+        }
+
+        @Override
+        public boolean isVisible(EventType t) {
+            return t.isVisible();
+        }
+    }
 
     // Used to provide EventSettings for jdk.management.jfr module
     static class DelegatedEventSettings extends EventSettings {
